+++ packages/rrweb/plugins/rrweb-plugin-sequential-id-record/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rrweb/rrweb-plugin-sequential-id-record
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`3d93020`](https://github.com/PostHog/posthog-js/commit/3d93020ead535e16c04d92e29c71a0805c5d760f)]:
+    - @posthog/rrweb@0.0.64
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @posthog/rrweb@0.0.63
+
 ## 0.0.1
 
 ### Patch Changes